requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
amqp
appdirs
asgiref
astunparse
attrs
beautifulsoup4
billiard
boto3
botocore
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cron-descriptor
Django
django-celery-beat
django-celery-results
django-debug-toolbar
django-timezone-field
djangorestframework
filelock
flatbuffers
frozendict
frozenlist
fsspec
gast
google-pasta
grpcio
h11
h5py
html5lib
huggingface-hub
idna
jmespath
joblib
keras
kombu
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
multidict
multitasking
namex
numpy
opt_einsum
optree
outcome
packaging
pandas
pandas_ta
patsy
peewee
plotly
prompt_toolkit
propcache
protobuf
pycparser
Pygments
PySocks
python-crontab
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
rich
s3transfer
safetensors
scikit-learn
scipy
selenium
setuptools
six
sniffio
sortedcontainers
soupsieve
sqlparse
statsmodels
ta
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tokenizers
tqdm
transformers
trio
trio-websocket
typing_extensions
tzdata
urllib3
vine
wcwidth
webencodings
websocket-client
Werkzeug
wheel
wrapt
wsproto
yarl
yfinance
